Moved Flask-Limiter code to a separate file. Made limit to be host based(global) instead of endpoint based.
